refactor(process): extract termination helpers to _termination.py
Move decide_termination_action (pure decision function) and
execute_termination_action (sole authorized async kill executor) from
process/__init__.py (942 → 821 lines) to process/_termination.py. The
public process.__init__ facade re-exports both so existing callers
keep working; _EXPECTED_PROCESS_SYMBOLS test continues to pass.

Refs: #4664
